Split tests into 2 different files
